requirements.txt
# Core ML Framework
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentence-transformers>=2.2.0

# Vector Database
faiss-cpu>=1.7.4

# Web Tools
duckduckgo-search>=3.8.0
trafilatura>=1.6.0
httpx>=0.24.0
requests>=2.31.0

# Configuration & Data
pyyaml>=6.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.82.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.4.0

# Utilities
python-dotenv>=1.0.0
